feat(read_child): wrapper function for knitr::knit_child
Thanks to @chantelwetzel-noaa for the example in your sa4ss documentation,
which inspired me to create this function. Now users won't have to
write three lines of code and won't botch the cat() call.
I tried to add as much documentation as I could, so now we can just
use the following code your documentation
example(sa4ss::read_child)
and the example text will be inserted. Let me know if this doesn't work.
We can also point users to the help file with help(sa4ss::read_child)

I might need to change the comment character in the example to be something
that works to suppress markdown rather than suppress R. Not sure.