fix(datetime2): wrongly calling datetime2-english instead of [english]{datetime2}
part of #7 where the package wasn't automatically installed b/c I was
calling it with the module as part of the name rather than in brackets.
Need this to set the language of the pdf.

There are still warning messages that can only be removed with
tlmgr_install("datetime2-english") and
tlmgr_install("glossaries-english") if you are using tinytex.