fix(sadraft.tex): accommodate pandoc 2.13
Thank you to all of the helpful github repositories that provided clear
commit messages when they upgraded their templates to accommodate the
changes to Pandoc from 2.8 to 2.13, e.g.,
https://github.com/FredHutch/VISCtemplates/pull/64
https://github.com/rstudio/rticles/commit/dec8eadaf4136f019dd6543c75d86af7e40c1598
The code actually comes from the
[Pandoc template](https://github.com/jgm/pandoc/blob/master/data/templates/default.latex).

The breaking change is from cslreferences to CSLReferences, along with some
other lines of code.

This fix was made in an attempt to accommodate previous versions of Pandoc as
well as future versions, and thus, the change in theory does not make sa4ss
dependent on any one specific version of Pandoc because knitting is already
a large enough pain in the kiester that we don't need more dependencies.

Close #12 with including pandoc instructions.

fix