feat: remove .Rmd file using _bookdown.yml
Helper file to remove the .Rmd file of the compiled document so
you don't get the stupid message of the file being present when
a run fails. This is very beta b/c I am not sure when the file
is actually deleted, but it seemed to be helpful to use the switch
provided by bookdown rather than writing my own code to do it.
See the bookdown book chapter on
configuration files](https://bookdown.org/yihui/bookdown/configuration.html)
for where I got the code.