bookdown.yaml was wrong saved entire folder instead of just log file
I don't see the log file in the folder, so I am going back to
saving just the folder with the three files, including the pdf.
I will need to work out how to save the log file.

Note that the path /home/runner/work/sa4ss/sa4ss/ gives access to all files
and it is unnecessary to upload a copy of the repository.