chore: Removes auto-pkg-maintenance
b/c the workflow was not up to date with the best practices
of GitHub actions and there is a new action in ghactions4r
that will format the DESCRIPTION file for you and that was
the main reason I was using this workflow.
Though grkstyle::grk_reindent_auto_pkg() was helpful.