fix: Removes nested acronyms in long entries
\newacronym{}{}{\gls{}} was causing acronyms to be used or counted
in the table of contents and future calls to the acronym were not
rendering the full text. By removing the use of the nested structure
and just typing the full name in the glossaries file it removes the
problem. The new problem is that the full name will always be used
in the first instance of the acronym even if a portion of the long name
has been previously referenced.

Close #101
Part of #21