fix: Use blue for links if enabled
Remove deprecated/duplicated code in sa4ss.sty that was being
overwritten with the information in the sa4ss template that is called
when compiling the file. This uses information in the yaml header
to turn things on. The default colors for all types of links are now
blue, that is just regular blue as defined by latex so we do not have
to call the color package.

Thanks to @brianlangseth-NOAA for wanting to make colored links, which
made me look at this code.