fix(margins): Sets all margins to 2.54 cm
Had to remove textwidth as @brianlangseth-NOAA so kindly noted was
leading to an error
```
Warning: Package geometry Warning: Over-specification in `h'-direction.
Warning:     `width' (398.33858pt) is ignored.
```
because geometry is hierarchical where the new input of margin conflicted
with the old input of overall textwidth so it was ignoring textwidth.
I have since removed the top and bottom margin calls as well as
textwidth in favor of setting the margins to 2.54 cm, i.e., 1 inch.

This change will NOT be automatically implemented in your sa4ss.sty file.
For those that have already ran draft(), you will need to change your
.sty file manually.

Close #104