Actually fixes the label problem of #tab:
I figured out that the weird label was only happening when we
called sa4ss::es_table_tex from inside of an .Rmd file. The bookdown
packages causes knitr to do weird things. I make R pretend that it is
not in bookdown because we are saving all the files externally and then
set the setting back to whatever it was using on.exit.

So, @chantelwetzel-noaa i am not sure if you brought the other branch
fix-escaping into your local branch but if you did you will want to
undo that commit or rebase back to main because I am going to bring
this change into main given it actually fixes the problem and does not
mess up people's escaping.