feat(add_figure): return the label from add_figure()
The label can then be used in report_ref_label() to make a label reference
in the document. I often get tired of remembering the label for a range
of figures. This will make the references of a string of figures much
easier. Though, you may have to render your document multiple times.